// src/features/core-stock/utils/enums.tsx
export enum SeriesFunction {
  INTRADAY = "intraday",
  TIME_SERIES_DAILY = "daily",
  TIME_SERIES_DAILY_ADJUSTED = "daily-adjusted",
  TIME_SERIES_WEEKLY = "TIME_SERIES_WEEKLY",
  TIME_SERIES_WEEKLY_ADJUSTED = "weekly-adjusted",
  TIME_SERIES_MONTHLY = "monthly",
  TIME_SERIES_MONTHLY_ADJUSTED = "monthly-adjusted",
}

export enum TimeSeriesByUrlParam {
  "intraday" = "TIME_SERIES_INTRADAY",
  "daily" = "TIME_SERIES_DAILY",
  "daily-adjusted" = "TIME_SERIES_DAILY_ADJUSTED",
  "weekly" = "TIME_SERIES_WEEKLY",
  "weekly-adjusted" = "TIME_SERIES_WEEKLY_ADJUSTED",
  "monthly" = "TIME_SERIES_MONTHLY",
  "monthly-adjusted" = "TIME_SERIES_MONTHLY_ADJUSTED",
}

export enum StockInformationKeys {
  INFORMATION = "1. Information",
  SYMBOL = "2. Symbol",
  LAST_REFRESHED = "3. Last Refreshed",
  INTERVAL = "4. Interval",
  OUTPUT_SIZE = "5. Output Size",
  TIME_ZONE = "6. Time Zone",
}

export enum Titles{
  intraday = "Intraday",
  daily = "Daily",
  "daily-adjusted" = "Daily Adjusted",
  weekly = "Weekly",
  "weekly-adjusted" = "Weekly Adjusted",
  monthly = "Monthly",
  "monthly-adjusted" = "Monthly Adjusted",
}

export enum TimeSeriesKeys {
  TIME_SERIES_INTRADAY = 'TIME_SERIES_INTRADAY',
  TIME_SERIES_DAILY = 'TIME_SERIES_DAILY',
  TIME_SERIES_DAILY_ADJUSTED = 'TIME_SERIES_DAILY_ADJUSTED',
  TIME_SERIES_WEEKLY = 'TIME_SERIES_WEEKLY',
  TIME_SERIES_WEEKLY_ADJUSTED = 'TIME_SERIES_WEEKLY_ADJUSTED',
  TIME_SERIES_MONTHLY = 'TIME_SERIES_MONTHLY',
  TIME_SERIES_MONTHLY_ADJUSTED = 'TIME_SERIES_MONTHLY_ADJUSTED',
}